Python 3 (Console) Projects
HANGMAN.rar is a version of the game played when you're bored called "Hangman"

MUSIC_THEORY_APP.rar is a console application that I created to test your memory and ability to recall Music Theory terms from ABRSM grades 3-5.

SCRABBLE2.0.rar is variation of the game called "Scrabble", it has a function which allows the computer to play.